requirements.txt
certifi==2019.9.11
cffi==1.13.1
cryptography==2.8
future==0.18.2
pycparser==2.19
python-telegram-bot==11.1.0
six==1.12.0
